Move helmChartConfig struct definition to config.go
